refactor: extract dtao conviction lock helpers and add unit tests
- Move conviction lock fetching into convictionLocks.ts, matching the
  module convention (alphaPrice, calculatePendingRootClaimable)
- Document the per-hotkey lock attribution approximation: on-chain the
  lock constrains the coldkey's total alpha on the subnet across all
  hotkeys, while we attribute it to the locked hotkey's token
- Add unit tests for toBigIntValue, u64f64RawToPlanck and the
  candidate/pair dedup helpers
